requirements.txt
numpy==1.24.3
Pillow==9.5.0
streamlit==1.27.1
tensorflow==2.13.0
keras==2.13.0